backend/requirements.txt
fastapi==0.63.0  # version fixed due to complete re-overriding fastapi objects
uvicorn
asyncpg
python-dotenv~=0.18.0
starlette_context

pyjwt
pydantic
passlib
argon2-cffi # required for passlib.hash.argon2

aiosmtplib # smtp

starlette-session
authlib
httpx

google-auth
google-auth-oauthlib
google-api-python-client

fastapi_utils

line-bot-sdk
